Move `DefaultIndexData` in its own file.
