+++ src/writer/item.cpp
@@ -19,8 +19,7 @@
 
 #include <zim/writer/item.h>
 #include <zim/writer/contentProvider.h>
-#include "xapian/myhtmlparse.h"
-#include "tools.h"
+#include "defaultIndexData.h"
 
 #include <sstream>
 
@@ -28,77 +27,6 @@ namespace zim
 {
   namespace writer
   {
-    class DefaultIndexData : public IndexData {
-      public:
-        DefaultIndexData(const std::string& htmlData, const std::string& title)
-          : title(title)
-        {
-#if defined(ENABLE_XAPIAN)
-          try {
-            htmlParser.parse_html(htmlData, "UTF-8", true);
-          } catch(...) {}
-#endif
-        }
-
-        bool hasIndexData() const {
-#if defined(ENABLE_XAPIAN)
-          return (htmlParser.dump.find("NOINDEX") == std::string::npos);
-#else
-          return false;
-#endif
-        }
-
-        std::string getTitle() const {
-#if defined(ENABLE_XAPIAN)
-          return zim::removeAccents(title);
-#else
-          return "";
-#endif
-         }
-
-        std::string getContent() const {
-#if defined(ENABLE_XAPIAN)
-          return zim::removeAccents(htmlParser.dump);
-#else
-          return "";
-#endif
-        }
-
-        std::string getKeywords() const {
-#if defined(ENABLE_XAPIAN)
-          return zim::removeAccents(htmlParser.keywords);
-#else
-          return "";
-#endif
-        }
-
-        uint32_t getWordCount() const {
-#if defined(ENABLE_XAPIAN)
-          return countWords(htmlParser.dump);
-#else
-          return 0;
-#endif
-        }
-
-        std::tuple<bool, double, double> getGeoPosition() const
-        {
-#if defined(ENABLE_XAPIAN)
-          if(htmlParser.has_geoPosition) {
-            return std::make_tuple(true, htmlParser.latitude, htmlParser.longitude);
-          }
-#endif
-          return std::make_tuple(false, 0, 0);
-        }
-
-      private:
-#if defined(ENABLE_XAPIAN)
-        zim::MyHtmlParser htmlParser;
-#endif
-        std::string title;
-
-    };
-
-
     std::unique_ptr<IndexData> Item::getIndexData() const
     {
       auto provider = getContentProvider();
